Reorganize init command
