+++ src/commands/init/params.ts
@@ -2,10 +2,10 @@ import defaultAvatar from "../../assets/defaultAvatar.js";
 import { releasesRecordFileName } from "../../utils/releaseRecord.js";
 
 export const stringsToRemoveFromName = [
-    "DAppNode-package-",
-    "DAppNodePackage-",
-    "DAppNode-Package",
-    "DAppNodePackage"
+  "DAppNode-package-",
+  "DAppNodePackage-",
+  "DAppNode-Package",
+  "DAppNodePackage"
 ];
 
 // Manifest
@@ -13,11 +13,11 @@ export const publicRepoDomain = ".public.dappnode.eth";
 export const defaultVersion = "0.1.0";
 
 // Avatar
-export const avatarPath = "avatar-default.png";
+export const avatarName = "avatar-default.png";
 export const avatarData = defaultAvatar;
 
 // Dockerfile
-export const dockerfilePath = "Dockerfile";
+export const dockerfileName = "Dockerfile";
 export const dockerfileData = `FROM busybox
   
   WORKDIR /usr/src/app
@@ -27,7 +27,7 @@ export const dockerfileData = `FROM busybox
   `;
 
 // .gitignore
-export const gitignorePath = ".gitignore";
+export const gitignoreName = ".gitignore";
 export const gitignoreCheck = "build_*";
 export const gitignoreData = `# DAppNodeSDK release directories
   build_*
