+++ test/commands/build.test.ts
@@ -1,6 +1,6 @@
 import { expect } from "chai";
 import { cleanTestDir, testDir } from "../testUtils.js";
-import { initHandler } from "../../src/commands/init/index.js";
+import { initHandler } from "../../src/commands/init/handler.js";
 import { buildHandler } from "../../src/commands/build.js";
 
 const contentProvider = "http://api.ipfs.dappnode.io:5001";
